requirements.txt
# Core ML Libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Data Processing
pyarrow==12.0.1
openpyxl==3.1.2

# API Framework
fastapi==0.101.1
uvicorn[standard]==0.23.2
pydantic==2.1.1

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
sqlite3

# Experiment Tracking & Model Registry
mlflow==2.5.0
optuna==3.2.0

# Monitoring & Observability
prometheus-client==0.17.1
grafana-api==1.0.3
evidently==0.4.2

# Data Validation
great-expectations==0.17.10
pandera==0.15.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Utilities
python-dotenv==1.0.0
click==8.1.6
tqdm==4.65.0
joblib==1.3.1

# Logging & Configuration
pyyaml==6.0.1
python-json-logger==2.0.7

# Containerization
docker==6.1.3

# Development
jupyter==1.0.0
ipykernel==6.24.0
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# HTTP Requests
requests==2.31.0
httpx==0.24.1

# Scheduling
apscheduler==3.10.1

# Security
cryptography==41.0.3
